+++ cmd/odrive-database/schema/functions.create.sql
@@ -3,4 +3,5 @@ source function.bitwise256_xor.sql;
 source function.keys.sql;
 source procedure.migrate_keys.sql;
 source procedure.rotate_keys.sql;
+source function.aacflatten.sql;